+++ app/models/cart_item/coupon.rb
@@ -2,12 +2,14 @@
 
 # A discount coupon applied to the whole shopping cart
 class CartItem::Coupon
+  attr_reader :errors
 
   # @param coupon {String|Coupon} may be nil or empty string if no coupons are applied
   def initialize(customer, operator, coupon)
     @customer = customer
     @operator = operator
     @coupon = coupon
+    @errors = {}
   end
 
   def coupon
@@ -26,5 +28,16 @@ class CartItem::Coupon
 
   def type
     'coupon'
+  end
+
+  def valid?(_all_items)
+    return true if @coupon.nil?
+
+    c = ::Coupon.find_by(code: @coupon)
+    if c.nil? || c.status(@customer.id) != 'active'
+      @errors[:item] = 'coupon is invalid'
+      return false
+    end
+    true
   end
 end